// src/types/profile.ts
export interface UserProfile {
  id: string;
  user_id: string;
  first_name: string;
  last_name: string;
  nationality: string;
  languages: string;
  marital_status: string;
  date_of_birth: string;
  phone_number: string;
  emergency_contact: string;
  address: string;
  created_at: string;
  updated_at: string;
}